--- 123.py
@@ -1,1 +1,0 @@
-print(1)
